fix for netsuite
