+++ packages/netsuite-adapter/src/adapter_creator.ts
@@ -14,22 +14,16 @@
 * limitations under the License.
 */
 import { Adapter, BuiltinTypes, ElemID, InstanceElement, ObjectType, AdapterInstallResult, AdapterOperationsContext, AdapterOperations } from '@salto-io/adapter-api'
-import { regex } from '@salto-io/lowerdash'
-import { logger } from '@salto-io/logging'
-import _ from 'lodash'
 import { SdkDownloadService } from '@salto-io/suitecloud-cli'
 import Bottleneck from 'bottleneck'
-import { CONFIG, configType, DEFAULT_CONCURRENCY, instanceLimiterCreator, NetsuiteConfig, validateClientConfig, validateDeployParams, validateFetchConfig, validateSuiteAppClientParams } from './config'
+import { configType, DEFAULT_CONCURRENCY, instanceLimiterCreator, netsuiteConfigFromConfig } from './config'
 import { NETSUITE } from './constants'
-import { validateFetchParameters, convertToQueryParams, validateNetsuiteQueryParameters, validateArrayOfStrings, validatePlainObject, FETCH_PARAMS } from './query'
 import { Credentials, isSdfCredentialsOnly, isSuiteAppCredentials, toCredentialsAccountId } from './client/credentials'
 import SuiteAppClient from './client/suiteapp_client/suiteapp_client'
 import SdfClient from './client/sdf_client'
 import NetsuiteClient from './client/client'
 import NetsuiteAdapter from './adapter'
 import loadElementsFromFolder from './sdf_folder_loader'
-
-const log = logger(module)
 
 const configID = new ElemID(NETSUITE)
 
@@ -71,95 +65,6 @@ export const defaultCredentialsType = new ObjectType({
   annotationRefsOrTypes: {},
   annotations: {},
 })
-
-const validateRegularExpressions = (regularExpressions: string[]): void => {
-  const invalidRegularExpressions = regularExpressions
-    .filter(strRegex => !regex.isValidRegex(strRegex))
-  if (!_.isEmpty(invalidRegularExpressions)) {
-    const errMessage = `received an invalid ${CONFIG.filePathRegexSkipList} value. The following regular expressions are invalid: ${invalidRegularExpressions}`
-    throw new Error(errMessage)
-  }
-}
-
-function validateConfig(config: Record<string, unknown>): asserts config is NetsuiteConfig {
-  const {
-    fetch,
-    fetchTarget,
-    skipList, // support deprecated version
-    deploy,
-    client,
-    filePathRegexSkipList,
-    typesToSkip,
-    suiteAppClient,
-  } = _.pick(config, Object.values(CONFIG))
-
-  if (filePathRegexSkipList !== undefined) {
-    validateArrayOfStrings(filePathRegexSkipList, CONFIG.filePathRegexSkipList)
-    validateRegularExpressions(filePathRegexSkipList)
-  }
-  if (typesToSkip !== undefined) {
-    validateArrayOfStrings(typesToSkip, CONFIG.typesToSkip)
-  }
-
-  if (client !== undefined) {
-    validatePlainObject(client, CONFIG.client)
-    validateClientConfig(client, fetchTarget !== undefined)
-  }
-
-  if (fetchTarget !== undefined) {
-    validatePlainObject(fetchTarget, CONFIG.fetchTarget)
-    validateNetsuiteQueryParameters(fetchTarget, CONFIG.fetchTarget)
-    validateFetchParameters(convertToQueryParams(fetchTarget))
-  }
-
-  if (skipList !== undefined) {
-    validatePlainObject(skipList, CONFIG.skipList)
-    validateNetsuiteQueryParameters(skipList, CONFIG.skipList)
-    validateFetchParameters(convertToQueryParams(skipList))
-  }
-
-  if (fetch !== undefined) {
-    validatePlainObject(fetch, CONFIG.fetch)
-    validateFetchConfig(fetch)
-  }
-
-  if (deploy !== undefined) {
-    validatePlainObject(deploy, CONFIG.deploy)
-    validateDeployParams(deploy)
-  }
-
-  if (suiteAppClient !== undefined) {
-    validatePlainObject(suiteAppClient, CONFIG.suiteAppClient)
-    validateSuiteAppClientParams(suiteAppClient)
-  }
-}
-
-const netsuiteConfigFromConfig = (
-  configInstance: Readonly<InstanceElement> | undefined
-): NetsuiteConfig => {
-  try {
-    if (!configInstance) {
-      return {}
-    }
-    const { value: config } = configInstance
-    validateConfig(config)
-    log.debug('using netsuite adapter config: %o', {
-      ...config,
-      fetch: _.omit(config.fetch, FETCH_PARAMS.lockedElementsToExclude),
-    })
-    return _.pickBy(config, (_value, key) => {
-      if (key in CONFIG) {
-        return true
-      }
-      log.debug('Unknown config property was found: %s', key)
-      return false
-    })
-  } catch (e) {
-    e.message = `Failed to load Netsuite config: ${e.message}`
-    log.error(e.message)
-    throw e
-  }
-}
 
 const throwOnMissingSuiteAppLoginCreds = (credentials: Credentials): void => {
   if (isSdfCredentialsOnly(credentials)) {
